requirements.txt
africastalking==1.2.7
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.5.1
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
djangorestframework==3.15.1
ecdsa==0.19.0
idna==3.7
oauthlib==3.2.2
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
schema==0.7.7
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
